@@ -6,7 +6,7 @@
     "node": "18.x"
   },
   "scripts": {
-    "dev": "nuxt",
+    "dev": "nuxt dev",
     "build": "nuxt build",
     "start": "nuxt start",
     "export": "nuxt export",
@@ -19,25 +19,19 @@
     "cypress:open": "npx cypress open"
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.13.6",
-    "@nuxtjs/i18n": "^7.3.1",
-    "@nuxtjs/pwa": "^3.3.5",
-    "@nuxtjs/sentry": "^8.0.7",
+    "@nuxt/eslint": "^1.2.0",
+    "@nuxtjs/i18n": "^9.3.1",
     "date-fns": "^2.30.0",
+    "eslint": "^9.22.0",
     "node-fetch-native": "^1.6.4",
-    "nuxt": "^2.18.1",
+    "nuxt": "^3.16.0",
     "tachyons": "^4.12.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.19.1",
-    "@nuxtjs/eslint-config": "^8.0.0",
-    "@nuxtjs/eslint-module": "^3.1.0",
-    "@nuxtjs/stylelint-module": "^4.1.0",
-    "cypress": "^13.11.0",
-    "eslint": "^8.3.0",
+    "@nuxtjs/eslint-module": "^4.1.0",
+    "@nuxtjs/stylelint-module": "^5.2.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-nuxt": "^2.0.0",
     "eslint-plugin-prettier": "^3.4.1",
     "postcss-html": "^1.5.0",
     "prettier": "^2.8.3",
