@@ -28,26 +28,23 @@
   </div>
 </template>
 
-<script>
-export default {
-  props: {
-    weather: {
-      type: Object,
-      default: () => ({
-        temperature: {
-          value: 'N/A',
-          unit: 'N/A',
-          place: 'N/A',
-        },
-        weatherIcons: [],
-        warningMessages: [],
-      }),
-    },
+<script setup>
+defineProps({
+  weather: {
+    type: Object,
+    default: () => ({
+      temperature: {
+        value: 'N/A',
+        unit: 'N/A',
+        place: 'N/A',
+      },
+      weatherIcons: [],
+      warningMessages: [],
+    }),
   },
-  methods: {
-    getImage(icon) {
-      return `https://www.hko.gov.hk/images/HKOWxIconOutline/pic${icon}.png`
-    },
-  },
+})
+
+function getImage(icon) {
+  return `https://www.hko.gov.hk/images/HKOWxIconOutline/pic${icon}.png`
 }
 </script>
