@@ -6,32 +6,28 @@
     <div class="pv1 tr">
       <div class="f7">
         {{ $t('last_updated') }}:
-        {{ updateDateTime | formatISODate }}
+        {{ formatISODate(updateDateTime, locale.code) }}
       </div>
     </div>
     <div class="pv2">
-      <district-weather :id="id" :districts="districts"></district-weather>
+      <DistrictWeather :id="id" :districts="districts" />
     </div>
   </div>
 </template>
 
-<script>
+<script setup>
 import { baseApiUrl } from '../variables.js'
+import { formatISODate } from '@/utils/formatUtils'
 
-export default {
-  async asyncData({ $axios, app }) {
-    const resWeather = await $axios.get(
-      baseApiUrl + `?dataType=rhrread&lang=${app.i18n.locale}`
-    )
+const { locale } = useI18n()
+const districts = ref([])
+const updateDateTime = ref(null)
+const id = ref('district-weather')
 
-    return {
-      id: 'district-weather',
-      districts: resWeather.data.temperature.data,
-      updateDateTime: resWeather.data.temperature.recordTime,
-    }
-  },
-  data() {
-    return {}
-  },
-}
+const { data: resWeather } = await useFetch(
+  baseApiUrl + `?dataType=rhrread&lang=${locale.value}`
+)
+
+districts.value = resWeather.value.temperature.data
+updateDateTime.value = resWeather.value.temperature.recordTime
 </script>
